pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loracode"
version = "0.1.2"
description = "AI-powered coding assistant in your terminal"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Lora Technologies", email = "support@loratech.dev"}
]
maintainers = [
    {name = "Lora Technologies", email = "support@loratech.dev"}
]
keywords = [
    "ai",
    "coding",
    "assistant",
    "terminal",
    "cli",
    "llm",
    "gpt",
    "developer-tools",
    "code-generation",
    "pair-programming"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Utilities",
]

dependencies = [
    "pydub",
    "configargparse",
    "GitPython",
    "jsonschema",
    "rich",
    "prompt_toolkit",
    "backoff",
    "pathspec",
    "diskcache",
    "grep_ast",
    "packaging",
    "sounddevice",
    "soundfile",
    "beautifulsoup4",
    "PyYAML",
    "diff-match-patch",
    "pypandoc",
    "litellm",
    "flake8",
    "importlib_resources",
    "pyperclip",
    "posthog",
    "mixpanel",
    "pexpect",
    "json5",
    "psutil",
    "watchfiles",
    "socksio",
    "pillow",
    "shtab",
    "oslex",
    "google-generativeai",
    "networkx<3.5",
    "scipy<1.16",
    "importlib-metadata<8.0.0",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "pytest-env",
    "pre-commit",
    "black",
    "isort",
    "mypy",
]
browser = [
    "streamlit",
]
help = [
    "playwright",
]

[project.scripts]
loracode = "loracode.main:main"

[project.urls]
Homepage = "https://github.com/Lora-Technologies/loracode"
Documentation = "https://github.com/Lora-Technologies/loracode#readme"
Repository = "https://github.com/Lora-Technologies/loracode.git"
Issues = "https://github.com/Lora-Technologies/loracode/issues"
Changelog = "https://github.com/Lora-Technologies/loracode/releases"

[tool.setuptools]
packages = ["loracode", "loracode.coders", "loracode.resources", "loracode.locales"]
include-package-data = true

[tool.setuptools.package-data]
loracode = [
    "resources/*",
    "queries/*",
    "locales/**/*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | venv
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".git", ".venv", "venv", "build", "dist"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
